requirements.txt
feedparser==6.0.11
groq==0.13.1
python-dotenv==1.0.1
requests==2.32.3
